requirements.txt
streamlit
python-dotenv
google-api-python-client
google-auth
langchain
langchain-community
requests